refactor: EditModal타입수정-86
